refactor: check spelling
