+++ demo/table.less
@@ -1,4 +1,4 @@
-table.pipline-table {
+table.pipeline-table {
   text-align: center;
 
   &,
