+++ packages/chart-advisor/src/advice-pipeline/data-to-dataprops.ts
